+++ packages/rhf-mui-components/src/global.d.ts
@@ -1,0 +1,2 @@
+declare module '*.css';
+declare module 'react-color-palette/css';
