+++ packages/rhf-mui-components/src/mui/file-uploader/index.tsx
@@ -22,7 +22,6 @@ import {
   type FormHelperTextProps
 } from '@/common';
 import { RHFMuiConfigContext } from '@/config/ConfigProvider';
-import type { FileInputProps, FileUploadError } from '@/types';
 import {
   fieldNameToLabel,
   keepLabelAboveFormField,
@@ -31,7 +30,11 @@ import {
 } from '@/utils';
 import { FileItem, HiddenInput, UploadButton } from './components';
 
-export type { FileUploadError };
+export enum FileUploadError {
+  sizeExceeded = 'FILE_SIZE_EXCEEDED',
+  invalidExtension = 'FILE_TYPE_NOT_ALLOWED',
+  limitExceeded = 'FILE_LIMIT_EXCEEDED',
+}
 
 export type RHFFileUploaderProps<T extends FieldValues> = {
   /**
@@ -50,6 +53,40 @@ export type RHFFileUploaderProps<T extends FieldValues> = {
    * When true, marks the field as required in the UI and accessibility attributes.
    */
   required?: boolean;
+  /**
+   * Comma-separated list of accepted file types.
+   *
+   * Examples:
+   * - `image/*`
+   * - `.pdf,.doc,.docx`
+   * - `image/png,image/jpeg`
+   */
+  accept?: string;
+  /**
+   * When true, allows selecting multiple files.
+   */
+  multiple?: boolean;
+  /**
+   * Maximum file size (in bytes) eligible for upload.
+   * Files exceeding this size will be rejected and trigger an error callback.
+   *
+   * For example, to set a maximum file size of 5 MB:
+   * `maxSize={5 * 1024 * 1024}`
+   */
+  maxSize?: number;
+  /**
+   * Maximum number of files that can be uploaded.
+   *
+   * When the limit is exceeded, additional files are rejected and
+   * reported through the error callback.
+   */
+  maxFiles?: number;
+  /**
+   * Disables the file input and prevents file selection.
+   *
+   * @default false
+   */
+  disabled?: boolean;
   /**
    * If true, displays each selected file size in the default file list.
    */
@@ -123,7 +160,7 @@ export type RHFFileUploaderProps<T extends FieldValues> = {
    * When true, the component expands to fill its container width.
    */
   fullWidth?: boolean;
-} & FileInputProps;
+};
 
 const RHFFileUploader = <T extends FieldValues>({
   fieldName,
