+++ packages/rhf-mui-components/src/types/index.ts
@@ -1,4 +1,3 @@
 export type * from './common';
 export type * from './config';
 export type * from './countries';
-export * from './file-uploader';
