+++ packages/rhf-mui-components/src/utils/file.ts
@@ -1,4 +1,4 @@
-import { FileUploadError } from '@/types';
+import { FileUploadError } from '@/mui/file-uploader';
 
 type FileSizeOptions = {
   valueAsNumber?: boolean;
